+++ open_manipulator_bringup/launch/omx_f_follower_ai.launch.py
@@ -1,0 +1,205 @@
+#!/usr/bin/env python3
+#
+# Copyright 2024 ROBOTIS CO., LTD.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Author: Wonho Yun, Sungho Woo, Woojin Wie, Junha Cha
+
+from launch import LaunchDescription
+from launch.actions import DeclareLaunchArgument
+from launch.actions import RegisterEventHandler
+from launch.conditions import IfCondition
+from launch.conditions import UnlessCondition
+from launch.event_handlers import OnProcessExit
+from launch.substitutions import Command
+from launch.substitutions import FindExecutable
+from launch.substitutions import LaunchConfiguration
+from launch.substitutions import PathJoinSubstitution
+from launch_ros.actions import Node
+from launch_ros.substitutions import FindPackageShare
+
+
+def generate_launch_description():
+    # Declare launch arguments
+    declared_arguments = [
+        DeclareLaunchArgument(
+            'start_rviz', default_value='false', description='Whether to execute rviz2'
+        ),
+        DeclareLaunchArgument(
+            'prefix',
+            default_value='""',
+            description='Prefix of the joint and link names',
+        ),
+        DeclareLaunchArgument(
+            'use_sim',
+            default_value='false',
+            description='Start robot in Gazebo simulation.',
+        ),
+        DeclareLaunchArgument(
+            'use_fake_hardware',
+            default_value='false',
+            description='Use fake hardware mirroring command.',
+        ),
+        DeclareLaunchArgument(
+            'fake_sensor_commands',
+            default_value='false',
+            description='Enable fake sensor commands.',
+        ),
+        DeclareLaunchArgument(
+            'init_position',
+            default_value='false',
+            description='Whether to launch the init_position node',
+        ),
+        DeclareLaunchArgument(
+            'ros2_control_type',
+            default_value='omx_f',
+            description='Type of ros2_control',
+        ),
+        DeclareLaunchArgument(
+            'port_name',
+            default_value='/dev/ttyACM0',
+            description='Port name for the hardware interface',
+        ),
+    ]
+
+    # Launch configurations
+    start_rviz = LaunchConfiguration('start_rviz')
+    prefix = LaunchConfiguration('prefix')
+    use_sim = LaunchConfiguration('use_sim')
+    use_fake_hardware = LaunchConfiguration('use_fake_hardware')
+    fake_sensor_commands = LaunchConfiguration('fake_sensor_commands')
+    init_position = LaunchConfiguration('init_position')
+    ros2_control_type = LaunchConfiguration('ros2_control_type')
+    port_name = LaunchConfiguration('port_name')
+
+    # Generate URDF file using xacro
+    urdf_file = Command([
+        PathJoinSubstitution([FindExecutable(name='xacro')]),
+        ' ',
+        PathJoinSubstitution([
+            FindPackageShare('open_manipulator_description'),
+            'urdf',
+            'omx_f',
+            'omx_f.urdf.xacro',
+        ]),
+        ' ',
+        'prefix:=',
+        prefix,
+        ' ',
+        'use_sim:=',
+        use_sim,
+        ' ',
+        'use_fake_hardware:=',
+        use_fake_hardware,
+        ' ',
+        'fake_sensor_commands:=',
+        fake_sensor_commands,
+        ' ',
+        'ros2_control_type:=',
+        ros2_control_type,
+        ' ',
+        'port_name:=',
+        port_name,
+    ])
+
+    # Paths for configuration files
+    controller_manager_config = PathJoinSubstitution([
+        FindPackageShare('open_manipulator_bringup'),
+        'config',
+        'omx_f_follower_ai',
+        'hardware_controller_manager.yaml',
+    ])
+
+    rviz_config_file = PathJoinSubstitution([
+        FindPackageShare('open_manipulator_description'),
+        'rviz',
+        'open_manipulator.rviz',
+    ])
+
+    trajectory_params_file = PathJoinSubstitution([
+        FindPackageShare('open_manipulator_bringup'),
+        'config',
+        'omx_f_follower_ai',
+        'initial_positions.yaml',
+    ])
+
+    # Define nodes
+    control_node = Node(
+        package='controller_manager',
+        executable='ros2_control_node',
+        parameters=[{'robot_description': urdf_file}, controller_manager_config],
+        output='both',
+        condition=UnlessCondition(use_sim),
+        remappings=[('/arm_controller/joint_trajectory', '/leader/joint_trajectory')],
+    )
+
+    robot_controller_spawner = Node(
+        package='controller_manager',
+        executable='spawner',
+        arguments=[
+            'arm_controller',
+            'joint_state_broadcaster',
+            # 'gpio_command_controller',
+        ],
+        output='both',
+        parameters=[{'robot_description': urdf_file}],
+    )
+
+    robot_state_publisher_node = Node(
+        package='robot_state_publisher',
+        executable='robot_state_publisher',
+        parameters=[{'robot_description': urdf_file, 'use_sim_time': use_sim}],
+        output='both',
+    )
+
+    joint_trajectory_executor = Node(
+        package='open_manipulator_bringup',
+        executable='joint_trajectory_executor',
+        parameters=[trajectory_params_file],
+        output='both',
+        condition=IfCondition(init_position),
+    )
+
+    rviz_node = Node(
+        package='rviz2',
+        executable='rviz2',
+        arguments=['-d', rviz_config_file],
+        output='both',
+        condition=IfCondition(start_rviz),
+    )
+
+    # Event handlers to ensure order of execution
+    delay_rviz_after_joint_state_broadcaster_spawner = RegisterEventHandler(
+        event_handler=OnProcessExit(
+            target_action=robot_controller_spawner, on_exit=[rviz_node]
+        )
+    )
+
+    delay_joint_trajectory_executor_after_controllers = RegisterEventHandler(
+        event_handler=OnProcessExit(
+            target_action=robot_controller_spawner,
+            on_exit=[joint_trajectory_executor],
+        )
+    )
+
+    return LaunchDescription(
+        declared_arguments
+        + [
+            control_node,
+            robot_controller_spawner,
+            robot_state_publisher_node,
+            delay_rviz_after_joint_state_broadcaster_spawner,
+            delay_joint_trajectory_executor_after_controllers,
+        ]
+    )