+++ open_manipulator_bringup/launch/omx_f_follower_ai_gazebo.launch.py
@@ -1,0 +1,168 @@
+#!/usr/bin/env python3
+#
+# Copyright 2024 ROBOTIS CO., LTD.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Author: Wonho Yun, Sungho Woo, Woojin Wie, Junha Cha
+
+import os
+from pathlib import Path
+
+from ament_index_python.packages import get_package_share_directory
+from launch import LaunchDescription
+from launch.actions import DeclareLaunchArgument, IncludeLaunchDescription
+from launch.actions import RegisterEventHandler, SetEnvironmentVariable
+from launch.event_handlers import OnProcessExit
+from launch.launch_description_sources import PythonLaunchDescriptionSource
+from launch.substitutions import Command, FindExecutable, LaunchConfiguration, PathJoinSubstitution
+from launch_ros.actions import Node
+from launch_ros.substitutions import FindPackageShare
+
+
+def generate_launch_description():
+    declared_arguments = [
+        DeclareLaunchArgument('model', default_value='omx_f',
+                              description='Robot model name.'),
+        DeclareLaunchArgument('world', default_value='empty_world',
+                              description='Gz sim World'),
+    ]
+
+    model = LaunchConfiguration('model')
+    world = LaunchConfiguration('world')
+
+    open_manipulator_description_path = os.path.join(
+        get_package_share_directory('open_manipulator_description'))
+
+    open_manipulator_bringup_path = os.path.join(
+        get_package_share_directory('open_manipulator_bringup'))
+
+    # Set gazebo sim resource path
+    gazebo_resource_path = SetEnvironmentVariable(
+        name='GZ_SIM_RESOURCE_PATH',
+        value=[
+            os.path.join(open_manipulator_bringup_path, 'worlds'), ':' +
+            str(Path(open_manipulator_description_path).parent.resolve())
+            ]
+        )
+
+    gazebo = IncludeLaunchDescription(
+                PythonLaunchDescriptionSource([os.path.join(
+                    get_package_share_directory('ros_gz_sim'), 'launch'), '/gz_sim.launch.py']),
+                launch_arguments=[
+                    ('gz_args', [
+                        world,
+                        '.sdf',
+                        ' -v 1',
+                        ' -r'
+                    ])
+                ]
+             )
+
+    robot_description_content = Command([
+        PathJoinSubstitution([FindExecutable(name='xacro')]),
+        ' ',
+        PathJoinSubstitution([FindPackageShare('open_manipulator_description'),
+                              'urdf',
+                              model,
+                              'omx_f.urdf.xacro']),
+        ' ',
+        'model:=', model,
+        ' ',
+        'use_sim:=true',
+        ' ',
+        'config_type:=omx_f_follower_ai'
+    ])
+
+    robot_description = {'robot_description': robot_description_content}
+
+    robot_state_pub_node = Node(
+        package='robot_state_publisher',
+        executable='robot_state_publisher',
+        parameters=[robot_description, {'use_sim_time': True}],
+        output='screen'
+    )
+
+    gz_spawn_entity = Node(
+        package='ros_gz_sim',
+        executable='create',
+        output='screen',
+        arguments=['-topic', 'robot_description',
+                   '-x', '0.0',
+                   '-y', '0.0',
+                   '-z', '0.0',
+                   '-R', '0.0',
+                   '-P', '0.0',
+                   '-Y', '0.0',
+                   '-name', model,
+                   '-allow_renaming', 'true',
+                   '-use_sim', 'true'],
+    )
+
+    joint_state_broadcaster_spawner = Node(
+        package='controller_manager',
+        executable='spawner',
+        arguments=['joint_state_broadcaster'],
+        output='screen'
+    )
+
+    robot_controller_spawner = Node(
+        package='controller_manager',
+        executable='spawner',
+        arguments=[
+            'arm_controller',
+            '--controller-ros-args',
+            '-r /arm_controller/joint_trajectory:=/leader/joint_trajectory',
+        ],
+        parameters=[robot_description],
+    )
+
+    bridge = Node(
+        package='ros_gz_bridge',
+        executable='parameter_bridge',
+        arguments=['/clock@rosgraph_msgs/msg/Clock[gz.msgs.Clock'],
+        output='screen'
+    )
+
+    # rviz_config_file = os.path.join(
+    #     open_manipulator_description_path, 'rviz', 'open_manipulator.rviz')
+
+    # rviz = Node(
+    #     package='rviz2',
+    #     executable='rviz2',
+    #     name='rviz2',
+    #     output='log',
+    #     arguments=['-d', rviz_config_file],
+    # )
+
+    return LaunchDescription([
+        *declared_arguments,
+        RegisterEventHandler(
+            event_handler=OnProcessExit(
+                target_action=gz_spawn_entity,
+                on_exit=[joint_state_broadcaster_spawner],
+            )
+        ),
+        RegisterEventHandler(
+            event_handler=OnProcessExit(
+               target_action=joint_state_broadcaster_spawner,
+               on_exit=[robot_controller_spawner],
+            )
+        ),
+        bridge,
+        gazebo_resource_path,
+        gazebo,
+        robot_state_pub_node,
+        gz_spawn_entity,
+        # rviz,
+    ])