+++ open_manipulator_bringup/launch/omy_f3m.launch.py
@@ -63,7 +63,7 @@ def generate_launch_description():
         ),
         DeclareLaunchArgument(
             'ros2_control_type',
-            default_value='omy_f3m',
+            default_value='omy_f3m_position',
             description='Type of ros2_control',
         ),
         DeclareLaunchArgument(